requirements.txt
python-dotenv
typer[all]
rich
duckdb
pandas
pyyaml
openpyxl
xlrd
pyxlsb
langchain
openai
pydantic
rich
pydantic_settings
